+++ src/MDNS_Generic.hpp
@@ -20,39 +20,30 @@
   You should have received a copy of the GNU Lesser General Public License along with EthernetBonjour.
   If not, see <http://www.gnu.org/licenses/>.
 
-  Version: 1.4.0
+  Version: 1.4.1
   
   Version  Modified By   Date      Comments
   -------  -----------  ---------- -----------
   1.0.0    K Hoang      01/08/2020 Initial coding to support W5x00 using Ethernet, EthernetLarge libraries
                                   Supported boards: nRF52, STM32, SAMD21/SAMD51, SAM DUE, Mega
-  1.0.1    K Hoang      02/10/2020 Add support to W5x00 using Ethernet2, Ethernet3 libraries
-  1.1.0    K Hoang      12/06/2021 Add support to RP2040-based boards
-  1.2.0    K Hoang      01/09/2021 Add support to generic boards using WiFi or WiFiNINA
-  1.2.1    K Hoang      02/09/2021 Remove support to ESP8266 to use native ESP8266mDNS library
-  1.3.0-b1 K Hoang      13/09/2021 Add support to Portenta_H7, using WiFi or Ethernet
+  ...
   1.3.0    K Hoang      28/09/2021 Add support to Portenta_H7, using WiFi or Ethernet
   1.3.1    K Hoang      10/10/2021 Update `platform.ini` and `library.json`
   1.4.0    K Hoang      26/01/2022 Fix `multiple-definitions` linker error
+  1.4.1    K Hoang      11/04/2022 Use Ethernet_Generic library as default. Support SPI1/SPI2 for RP2040/ESP32
  *****************************************************************************************************************************/
 
 #ifndef __MDNS_GENERIC_HPP__
 #define __MDNS_GENERIC_HPP__
 
-#if ( SYSTEM_ENDIAN == _ENDIAN_LITTLE_ )
-  #warning SYSTEM_ENDIAN == _ENDIAN_LITTLE_
-#else
-  #warning SYSTEM_ENDIAN == _ENDIAN_BIG_
-#endif
-
 #ifndef MDNS_GENERIC_VERSION
-  #define MDNS_GENERIC_VERSION            "MDNS_Generic v1.4.0"
+  #define MDNS_GENERIC_VERSION            "MDNS_Generic v1.4.1"
 
   #define MDNS_GENERIC_VERSION_MAJOR      1
   #define MDNS_GENERIC_VERSION_MINOR      4
-  #define MDNS_GENERIC_VERSION_PATCH      0
+  #define MDNS_GENERIC_VERSION_PATCH      1
 
-#define MDNS_GENERIC_VERSION_INT        1004000
+#define MDNS_GENERIC_VERSION_INT        1004001
 
 #endif
 
@@ -65,6 +56,13 @@ extern "C"
 
 #include "MDNS_Generic_Debug.h"
 
+#if (_MDNS_LOGLEVEL_ > 2)
+  #if ( SYSTEM_ENDIAN == _ENDIAN_LITTLE_ )
+    #warning SYSTEM_ENDIAN == _ENDIAN_LITTLE_
+  #else
+    #warning SYSTEM_ENDIAN == _ENDIAN_BIG_
+  #endif
+#endif
 
 typedef uint8_t byte;